Fix duration signal race conditions which leave hub page durationless
Before this commit duration signal is sent to the hub page as soon
as Youtube responds. Because of that behaviour the following sequences
of events were possible:
1.
    hub opens
    duration event A
    new video payload
    duration update transaction complete

2.
    duration event A
    hub opens
    new video payload
    duration update transaction complete

3.
    duration event A
    new video payload
    hub opens
    duration update transaction complete

In all of these cases the hub page is left not showing the duration
sent in event A, either because A was missed, or the new video payload
clobbers the duration in A.

This commit ensure that duration events fire only after the transaction
complete, which prevent these sequences from happening.