Make hub page work. I totally lied in the last commit
It looks like any page packaged in the add-on gets the treatment of
the options page. A downside is that this might go away in the future
since the Chrome docs says they don't recommand the options page API.

At the moment though, this setup allows for the best product.